explicit casting of 'long long' vs 'long' for windows compatibility.

+++ test/hotspot/gtest/utilities/test_globalDefinitions.cpp
@@ -304,8 +304,8 @@ TEST(globalDefinitions, jlong_from) {
   EXPECT_EQ(val, 0x00000000000000FF);
 
   val = jlong_from(0xFFFFFFFF, 0);
-  EXPECT_EQ(               val, (signed long)0xFFFFFFFF00000000);
-  EXPECT_EQ((unsigned long)val,              0xFFFFFFFF00000000);
+  EXPECT_EQ(                    val, (signed long long)0xFFFFFFFF00000000);
+  EXPECT_EQ((unsigned long long)val,                   0xFFFFFFFF00000000);
 
   val = jlong_from(0, 0xFFFFFFFF);
   EXPECT_EQ(val, 0x00000000FFFFFFFF);
@@ -314,10 +314,10 @@ TEST(globalDefinitions, jlong_from) {
   EXPECT_EQ(val, 0x00000000FFFFFFFF);
 
   val = jlong_from(-1, 0);
-  EXPECT_EQ(val, (signed long)0xFFFFFFFF00000000);
+  EXPECT_EQ(val, (signed long long)0xFFFFFFFF00000000);
 
   val = jlong_from(-1, -1);
-  EXPECT_EQ(val, (signed long)0xFFFFFFFFFFFFFFFF);
+  EXPECT_EQ(val, (signed long long)0xFFFFFFFFFFFFFFFF);
   EXPECT_EQ(val, -1);
 
   val = jlong_from(0xABCD, 0xEFEF);
